split menu.cpp/.h into files
