+++ CMakeLists.txt
@@ -22,6 +22,7 @@ add_executable(
         source/map.cpp
         source/controller.cpp
         source/menu.cpp
+        source/menu_items.cpp
         source/pause.cpp
         source/main_menu.cpp
         source/game_environment.cpp
